Seprate the execution logic
Extract the execution logic to a simple interface
Created a simple ExecutorFactory
Created a simple SerialExecutor by using the cli.apply_plan
Used SerialExecutor in cli
