+++ cou/cli.py
@@ -22,10 +22,8 @@ import sys
 from datetime import datetime
 from typing import Any
 
-from colorama import Fore, Style
-
-from cou.steps import UpgradeStep
 from cou.steps.analyze import Analysis
+from cou.steps.execute import ExecutorFactory
 from cou.steps.plan import generate_plan
 from cou.utils import juju_utils as utils
 
@@ -124,56 +122,6 @@ def setup_logging(log_level: str = "INFO") -> None:
     logger.info("Logs of this execution can be found at %s", file_name)
 
 
-def prompt(parameter: str) -> str:
-    """Generate eye-catching prompt.
-
-    :param parameter: String to show at the prompt with the user options.
-    :type parameter: str
-    :return: Colored prompt string with the user options.
-    :rtype: str
-    """
-
-    def bold(text: str) -> str:
-        return Style.RESET_ALL + Fore.RED + Style.BRIGHT + text + Style.RESET_ALL
-
-    def normal(text: str) -> str:
-        return Style.RESET_ALL + Fore.RED + text + Style.RESET_ALL
-
-    return (
-        normal(parameter + " (")
-        + bold("c")
-        + normal(")ontinue/(")
-        + bold("a")
-        + normal(")bort/(")
-        + bold("s")
-        + normal(")kip:")
-    )
-
-
-async def apply_plan(upgrade_plan: UpgradeStep, non_interactive: bool) -> None:
-    """Apply the plan for upgrade.
-
-    :param upgrade_plan: Plan to be executed on steps.
-    :type upgrade_plan: UpgradeStep
-    """
-    result = "X"
-    while result.casefold() not in AVAILABLE_OPTIONS:
-        result = "c" if non_interactive else input(prompt(upgrade_plan.description)).casefold()
-        match result:
-            case "c":
-                logger.info("Running: %s", upgrade_plan.description)
-                await upgrade_plan.run()
-                for sub_step in upgrade_plan.sub_steps:
-                    await apply_plan(sub_step, non_interactive)
-            case "a":
-                logger.info("Aborting plan")
-                sys.exit(1)
-            case "s":
-                logger.info("Skipped")
-            case _:
-                logger.info("No valid input provided!")
-
-
 async def entrypoint() -> None:
     """Execute 'charmed-openstack-upgrade' command."""
     try:
@@ -188,7 +136,7 @@ async def entrypoint() -> None:
         print(analysis_result)
         upgrade_plan = await generate_plan(analysis_result)
         if args.run:
-            await apply_plan(upgrade_plan, args.non_interactive)
+            await ExecutorFactory.create_executor(upgrade_plan, not args.non_interactive).execute()
         else:
             print(upgrade_plan)
